requirements.txt
anthropic>=0.40.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
librosa>=0.10.0
soundfile>=0.12.0
python-dotenv>=1.0.0
click>=8.1.0
pytest>=8.0.0
